Clean up SqlQueryPlanJoinBuilder.make_combine_metrics_join_description
Minor cleanup on the method plus some reorganization of helper functions.
In particular, the make_coalesced_expr has been moved to a separate file,
which might prove useful later on as we consolidate our time range expression
building or other things of that nature.
